Checks: Add check_postgres_pgbouncer_requests.py
This script checks the requests per second on the connection pooler of
postgresql called pgbouncer. It raises a warning or critical state if
a reasonable threshold is reached.
Values for the database to be monitored, the used port of pgbouncer and
the warning and critical thresholds can be specified using parameters.