FIX: check_systemd_units auto-restart handling
It's a common behavior that services have state activating and sub auto-restart.
To catch services that are in restart cycle due to an error,
we have to check the exitcode to identify problems.
Throwing warnings if a service is in state activating, does not make sense.
To catch "hanging" services we need some other method.