Allow for remote hosts checks
This version keeps the compatibility to use the pemfile either as
a positional argument or as a named one for historic reasons.
You are strongly advised to use --pemfile as the positional option
will be removed.

You can now also use --remote=(hostname|IP) to check an online host
and --name will allow you to set the hostname for SNI.