Ignore system service units if auto restarting
We don't want to raise a warning if a system service unit is in state
activating and sub state auto-restarting because systemd can probably
fix it on its own and there is no manual action required.
If this restart would fail we would notice later when the unit is in
state failed.