Introduce the check_bgp_protocols.py
This is a python3 replacement for the check_bgp_peers.sh.

check_bgp_peers.sh has hardcoded paths and can only check protocols for
bird IPv4.
This check is capable of checking either inet families, requiring just that
you provide the proper path to the birdc/birdc6 binary, meaning that for
checking IPv4 and IPv6, you will need to run the check twice.

It will return critical if a protocol is not up and established, warning
for protocols that were disabled on purpose and unknown if you use a
protocol that this check doesn't know how to handle yet.