Introduce the script to export VM data from HVs
We export data from all VMs found in libvirt on a given hypervisor to
the Redis servers. Each VM represent a key in redis with the
`igvm_consistency:<hv>:<vm>` format. The value is a set containing
object_id, libvirt state and libvirt reason.

Then the igvm_consistency_check will, together with the serveradmin
data, determine if VMs are in a valid state as well as find missing and
duplicated VMs.