lbpool_nodes: Warn if not all nodes are alive
The case where more than min_nodes but less than optimal_nodes are alive
was returning an OK state instead of Warning.

It was mistakenly defined like that to cover the case of clusters, but
at that level, a single HWLB group is being checked and therefore,
shouldn't be aware of the cluster logic.