check_interfaces: Force microseconds on timestamp
Python has an interesting behavior where if you call just `str()` on a
timestamp where the microseconds value is zero, it will drop the microseconds
completely.

So if you have something like `2025-12-03 09:47:51.000000`, when you
`str()`, you end up with `2025-12-03 09:47:51`, causing the
`datetime.strptime()` call in the `from_json()` function to fail as the
format does not match anymore.